Upload data 2021년 4월 10일 토요일 18시 13분 54초 KST

+++ make.sh
@@ -4,12 +4,19 @@ echo -e "\033[0;32mMake new One...\033[0m"
 
 num=$1
 
-# make new dir
-mkdir $num
- 
-# cd dir
-cd $num
+# make new di
+if test -d "./$num";then
+	echo "dir already exist"
+else
+	mkdir $num
+	# cd dir
+	cd $num
 
-# cp sample
-cp ../sample.cpp $num.cpp
-
+	# cp sample
+	if test -f "./$num.cpp";then
+		echo "file already exist"
+	else
+		cp ../sample.cpp $num.cpp
+		echo "make new file"
+	fi
+fi
